Adds translation performance to performance script
