+++ packages/graphql/tests/performance/databaseQuery/cypher/aggregation.cypher
