+++ packages/graphql/tests/performance/databaseQuery/cypher/basic.cypher
