+++ packages/graphql/tests/performance/databaseQuery/setup-database.ts
@@ -579,11 +579,11 @@ CREATE (user6)-[:LIKES]->(JohnnyMnemonic)
 CREATE (user6)-[:LIKES]->(CloudAtlas)
 
 // Dummy movies with one actor
-CREATE (m:Movie {name: "Sharknado", released: 2013})
+CREATE (m:Movie {title: "Sharknado", released: 2013})
 WITH *
 UNWIND range(0, 1000) AS x
 CREATE (p:Person {born: 1000+x, name: "Shark "+x})
-CREATE (m2:Movie {name: "Sharknado "+x, released: 2013})
+CREATE (m2:Movie {title: "Sharknado "+x, released: 2013})
 CREATE (p)-[:ACTED_IN]->(m)
 CREATE (p)-[:ACTED_IN]->(m2)
