+++ packages/graphql/tests/performance/schema/schema-performance.ts
@@ -17,11 +17,9 @@
  * limitations under the License.
  */
 
-import Neo4jGraphQL from "../../src/classes/Neo4jGraphQL";
+import Neo4jGraphQL from "../../../src/classes/Neo4jGraphQL";
 
 const basicTypeDefs = `
-    extend schema @link(url: "https://specs.apollo.dev/federation/v2.0", import: ["@key"])
-
     type Journalist {
         articles: [Article!]! @relationship(type: "HAS_ARTICLE", direction: OUT, properties: "HasArticle")
     }
@@ -30,8 +28,8 @@ const basicTypeDefs = `
         createdAt: DateTime! @timestamp
     }
 
-    type Article @key(fields: "id") {
-        id: ID! @id
+    type Article @authorization(filter: [{ where: { node: { id: "$jwt.sub" } } }]) {
+        id: ID! @id @authorization(filter: [{ where: { node: { id: "$jwt.sub" } } }])
         blocks: [Block!]! @relationship(type: "HAS_BLOCK", direction: OUT, properties: "HasBlock")
         images: [Image!]! @relationship(type: "HAS_IMAGE", direction: OUT)
     }
@@ -68,7 +66,7 @@ const basicTypeDefs = `
     }
 `;
 
-export async function subgraphSchemaPerformance() {
+export async function schemaPerformance() {
     let typeDefs = "";
     const toReplace =
         /(Journalist|Article|HasArticle|Block|Image|HasBlock|TextBlock|DividerBlock|ImageBlock|PDFImage|HAS_ARTICLE|HAS_BLOCK|HAS_IMAGE)/g;
@@ -80,7 +78,7 @@ export async function subgraphSchemaPerformance() {
     const neoSchema = new Neo4jGraphQL({
         typeDefs,
     });
-    console.time("Subgraph Schema Generation");
-    await neoSchema.getSubgraphSchema();
-    console.timeEnd("Subgraph Schema Generation");
+    console.time("Schema Generation");
+    await neoSchema.getSchema();
+    console.timeEnd("Schema Generation");
 }
