--- packages/graphql/tests/performance/translation/graphql/connect.graphql
@@ -1,13 +1,0 @@
-mutation createAndConnect {
-    createMovies(
-        input: { title: "My Movie", id: 5, actors: { connect: { where: { node: { name_CONTAINS: "Shark" } } } } }
-    ) {
-        movies {
-            title
-        }
-        info {
-            nodesCreated
-            relationshipsCreated
-        }
-    }
-}
