--- packages/graphql/tests/performance/translation/graphql/connections.graphql
@@ -1,65 +1,0 @@
-query Connection {
-    movies {
-        actorsConnection {
-            edges {
-                node {
-                    name
-                }
-            }
-        }
-    }
-}
-
-query NestedConnection {
-    movies {
-        actorsConnection {
-            edges {
-                node {
-                    name
-                    moviesConnection {
-                        edges {
-                            node {
-                                title
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-query ConnectionWithSort {
-    moviesConnection(first: 5, sort: { title: ASC }) {
-        edges {
-            node {
-                title
-                actorsConnection {
-                    edges {
-                        node {
-                            name
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-query ConnectionWithSortAndCypher {
-    moviesConnection(first: 5, sort: [{ title: ASC }, { oneActorName: DESC }]) {
-        edges {
-            node {
-                title
-                oneActorName
-                actorsConnection {
-                    edges {
-                        node {
-                            name
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
