--- packages/graphql/tests/performance/translation/graphql/query.graphql
@@ -1,28 +1,0 @@
-query SimpleQuery {
-    movies {
-        title
-    }
-}
-
-query SimpleAggregate {
-    moviesAggregate {
-        count
-    }
-}
-
-query DeeplyNestedConnectionWhere {
-    movies(
-        where: {
-            actorsConnection_SOME: {
-                node: {
-                    name: "Hugo Weaving"
-                    moviesConnection_NONE: {
-                        node: { title: "The Matrix", directorsConnection_SOME: { node: { name_CONTAINS: "Lana" } } }
-                    }
-                }
-            }
-        }
-    ) {
-        title
-    }
-}
